SetValor tava int, mudança pra double

+++ RP2T3/src/Imovel/Imovel.java
@@ -153,7 +153,7 @@ public abstract class Imovel {
     /**
      * @param valor the valor to set
      */
-    public void setValor(int valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
